requirements.txt
streamlit==1.38.0
langchain==0.2.14
langchain-huggingface==0.0.3
huggingface_hub==0.24.6
python-dotenv==1.0.1